Seperated each window into files
